refactor(agent-tars-cli): decouple 4
refactor(agent): move workspace type to agent kernel
